third_party/blink/web_tests/wpt_internal/css/css-flexbox/balance/flex-line-count-valid.html

<!DOCTYPE html>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/3070">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("flex-line-count", "1");
test_valid_value("flex-line-count", "2");
</script>
</body>
